Remove chart level getPosition and getAlteredPosition
