--- packages/polaris-viz/src/components/HorizontalBarChart/utilities/index.ts
@@ -1,1 +1,0 @@
-export {getAlteredHorizontalBarPosition} from './getAlteredHorizontalBarPosition';
