+++ packages/polaris-viz/src/components/LineChart/utilities/index.ts
@@ -1,2 +1,1 @@
 export {yAxisMinMax} from './yAxisMinMax';
-export {getAlteredLineChartPosition} from './getAlteredLineChartPosition';
