+++ packages/polaris-viz/src/components/StackedAreaChart/Chart.tsx
@@ -18,7 +18,7 @@ import {
   useChartPositions,
   LINE_HEIGHT,
   SMALL_CHART_HEIGHT,
-  clamp,
+  InternalChartType,
 } from '@shopify/polaris-viz-core';
 
 import {ChartElements} from '../ChartElements';
@@ -35,17 +35,7 @@ import type {
 } from '../../types';
 import {XAxis} from '../XAxis';
 import {LegendContainer, useLegend} from '../LegendContainer';
-import type {
-  TooltipPosition,
-  TooltipPositionOffset,
-  TooltipPositionParams,
-} from '../TooltipWrapper';
-import {
-  TooltipHorizontalOffset,
-  TooltipVerticalOffset,
-  TooltipWrapper,
-  TOOLTIP_POSITION_DEFAULT_RETURN,
-} from '../TooltipWrapper';
+import {TooltipWrapper} from '../TooltipWrapper';
 import {
   useLinearChartAnimations,
   useTheme,
@@ -54,7 +44,6 @@ import {
   useLinearLabelsAndDimensions,
 } from '../../hooks';
 import {ChartMargin, ANNOTATIONS_LABELS_OFFSET} from '../../constants';
-import {eventPointNative} from '../../utilities';
 import {YAxis} from '../YAxis';
 import {Crosshair} from '../Crosshair';
 import {VisuallyHiddenRows} from '../VisuallyHiddenRows';
@@ -64,13 +53,7 @@ import {useStackedData} from './hooks';
 import {StackedAreas, Points} from './components';
 import {useStackedChartTooltipContent} from './hooks/useStackedChartTooltipContent';
 import {yAxisMinMax} from './utilities/yAxisMinMax';
-import {getAlteredStackedAreaChartPosition} from './utilities/getAlteredStackedAreaChartPosition';
 import styles from './Chart.scss';
-
-const TOOLTIP_POSITION: TooltipPositionOffset = {
-  horizontal: TooltipHorizontalOffset.Left,
-  vertical: TooltipVerticalOffset.Center,
-};
 
 export interface Props {
   annotationsLookupTable: AnnotationLookupTable;
@@ -395,17 +378,18 @@ export function Chart({
 
       {longestSeriesLength !== -1 && (
         <TooltipWrapper
-          alwaysUpdatePosition
           chartBounds={chartBounds}
+          chartType={InternalChartType.Line}
+          data={data}
           focusElementDataType={DataType.Point}
           getMarkup={getTooltipMarkup}
-          getPosition={getTooltipPosition}
-          getAlteredPosition={getAlteredStackedAreaChartPosition}
           id={tooltipId}
+          longestSeriesIndex={longestSeriesIndex}
           margin={ChartMargin}
           onIndexChange={(index) => setActivePointIndex(index)}
           parentRef={svgRef}
           usePortal
+          xScale={xScale}
         />
       )}
 
@@ -422,47 +406,4 @@ export function Chart({
       )}
     </ChartElements.Div>
   );
-
-  function getTooltipPosition({
-    event,
-    index,
-    eventType,
-  }: TooltipPositionParams): TooltipPosition {
-    if (eventType === 'mouse' && event) {
-      const point = eventPointNative(event!);
-
-      if (point == null || xScale == null) {
-        return TOOLTIP_POSITION_DEFAULT_RETURN;
-      }
-
-      const {svgX} = point;
-
-      const closestIndex = Math.round(xScale.invert(svgX - chartXPosition));
-
-      const activeIndex = clamp({
-        amount: closestIndex,
-        min: 0,
-        max: data[longestSeriesIndex].data.length - 1,
-      });
-
-      return {
-        x: (event as MouseEvent).pageX,
-        y: (event as MouseEvent).pageY,
-        position: TOOLTIP_POSITION,
-        activeIndex,
-      };
-    } else if (index != null) {
-      const activeIndex = index ?? 0;
-      const x = xScale?.(activeIndex) ?? 0;
-
-      return {
-        x: x + (dimensions?.x ?? 0),
-        y: dimensions?.y ?? 0,
-        position: TOOLTIP_POSITION,
-        activeIndex: index,
-      };
-    }
-
-    return TOOLTIP_POSITION_DEFAULT_RETURN;
-  }
 }
