+++ packages/polaris-viz/src/components/TooltipWrapper/components/TooltipAnimatedContainer.tsx
@@ -1,14 +1,15 @@
 import type {ReactNode} from 'react';
 import {useEffect, useRef, useState, useMemo} from 'react';
 import type {BoundingRect, Dimensions} from '@shopify/polaris-viz-core';
-import {useChartContext} from '@shopify/polaris-viz-core';
+import {useChartContext, InternalChartType} from '@shopify/polaris-viz-core';
 
 import {isTouchDevice} from '../../../utilities/isTouchDevice';
 import type {Margin} from '../../../types';
 import type {TooltipPositionOffset} from '../types';
-import {getAlteredVerticalBarPosition} from '../utilities';
-import type {AlteredPosition} from '../utilities';
 import {DEFAULT_TOOLTIP_POSITION} from '../constants';
+import {getAlteredLineChartPosition} from '../utilities/getAlteredLineChartPosition';
+import {getAlteredHorizontalBarPosition} from '../utilities/getAlteredHorizontalBarPosition';
+import {getAlteredVerticalBarPosition} from '../utilities/getAlteredVerticalBarPosition';
 
 import styles from './TooltipAnimatedContainer.scss';
 
@@ -19,25 +20,23 @@ export interface TooltipAnimatedContainerProps {
   currentX: number;
   currentY: number;
   chartBounds: BoundingRect;
-  getAlteredPosition?: AlteredPosition;
+  chartType: InternalChartType;
   position?: TooltipPositionOffset;
   id?: string;
   bandwidth?: number;
-  chartDimensions?: BoundingRect;
 }
 
 export function TooltipAnimatedContainer({
   activePointIndex,
   bandwidth = 0,
   chartBounds,
+  chartType,
   children,
   currentX,
   currentY,
   id = '',
-  getAlteredPosition = getAlteredVerticalBarPosition,
   margin,
   position = DEFAULT_TOOLTIP_POSITION,
-  chartDimensions,
 }: TooltipAnimatedContainerProps) {
   const {isPerformanceImpacted, scrollContainer} = useChartContext();
 
@@ -46,13 +45,25 @@ export function TooltipAnimatedContainer({
     useState<Dimensions | null>(null);
   const firstRender = useRef(true);
 
+  const getAlteredPositionFunction = useMemo(() => {
+    switch (chartType) {
+      case InternalChartType.Line:
+        return getAlteredLineChartPosition;
+      case InternalChartType.HorizontalBar:
+        return getAlteredHorizontalBarPosition;
+      case InternalChartType.Bar:
+      default:
+        return getAlteredVerticalBarPosition;
+    }
+  }, [chartType]);
+
   const {x, y, opacity, immediate} = useMemo(() => {
     if (tooltipDimensions == null) {
       return {x: 0, y: 0, opacity: 0};
     }
     const touchDevice = isTouchDevice();
 
-    const {x, y} = getAlteredPosition({
+    const {x, y} = getAlteredPositionFunction({
       currentX,
       currentY,
       position,
@@ -62,7 +73,6 @@ export function TooltipAnimatedContainer({
       bandwidth,
       isPerformanceImpacted,
       isTouchDevice: touchDevice,
-      chartDimensions,
       scrollContainer,
     });
 
@@ -80,12 +90,11 @@ export function TooltipAnimatedContainer({
     chartBounds,
     currentX,
     currentY,
-    getAlteredPosition,
+    getAlteredPositionFunction,
     margin,
     position,
     isPerformanceImpacted,
     tooltipDimensions,
-    chartDimensions,
     scrollContainer,
   ]);
