+++ packages/polaris-viz/src/components/TooltipWrapper/utilities.ts
@@ -1,145 +1,5 @@
-import type {BoundingRect, Dimensions} from '@shopify/polaris-viz-core';
-import {clamp} from '@shopify/polaris-viz-core';
-
-import type {TooltipPositionOffset} from '../TooltipWrapper';
-import type {Margin} from '../../types';
-
-import {TooltipHorizontalOffset, TooltipVerticalOffset} from './types';
-
-// The space between the cursor and the tooltip
-export const TOOLTIP_MARGIN = 20;
-export const SCROLLBAR_WIDTH = 20;
-
-export interface AlteredPositionProps {
-  bandwidth: number;
-  chartBounds: BoundingRect;
-  currentX: number;
-  currentY: number;
-  isPerformanceImpacted: boolean;
-  isTouchDevice: boolean;
-  margin: Margin;
-  position: TooltipPositionOffset;
-  tooltipDimensions: Dimensions;
-  chartDimensions?: BoundingRect;
-  scrollContainer?: Element | null;
-}
-
-export interface AlteredPositionReturn {
-  x: number;
-  y: number;
-}
-
-export type AlteredPosition = (
-  props: AlteredPositionProps,
-) => AlteredPositionReturn;
-
-// Keep the tooltip within the bounds of the chart.
-// Based on "position" the tooltip will be placed
-// around the chart item so the item should never
-// be obscured by the tooltip.
-export function getAlteredVerticalBarPosition(
-  props: AlteredPositionProps,
-): AlteredPositionReturn {
-  const {currentX, currentY, position, scrollContainer} = props;
-
-  const newPosition = {...position};
-
-  let x = currentX;
-  let y = currentY - (scrollContainer?.scrollTop ?? 0);
-
-  //
-  // Y POSITIONING
-  //
-  if (!props.isPerformanceImpacted) {
-    if (newPosition.vertical === TooltipVerticalOffset.Inline) {
-      newPosition.horizontal = TooltipHorizontalOffset.Left;
-
-      const inline = getInlinePosition(y, props);
-      y = inline.value;
-    }
-
-    if (newPosition.vertical === TooltipVerticalOffset.Center) {
-      const verticalCenter = getVerticalCenterPosition(y, props);
-      y = verticalCenter.value;
-    }
-
-    if (newPosition.vertical === TooltipVerticalOffset.Above) {
-      const above = getAbovePosition(y, props);
-      y = above.value;
-
-      if (above.wasOutsideBounds) {
-        newPosition.horizontal = TooltipHorizontalOffset.Left;
-      }
-    }
-
-    if (newPosition.vertical === TooltipVerticalOffset.Below) {
-      const below = getBelowPosition(y, props);
-      y = below.value;
-
-      if (below.wasOutsideBounds) {
-        newPosition.horizontal = TooltipHorizontalOffset.Left;
-      }
-    }
-  } else {
-    y = clamp({
-      amount:
-        (props.chartDimensions?.y ?? 0) -
-        props.tooltipDimensions.height -
-        (scrollContainer?.scrollTop ?? 0),
-      min: 0,
-      max:
-        window.scrollY +
-        window.innerHeight -
-        props.tooltipDimensions.height -
-        TOOLTIP_MARGIN,
-    });
-  }
-
-  //
-  // X POSITIONING
-  //
-
-  if (newPosition.horizontal === TooltipHorizontalOffset.Left) {
-    const left = getLeftPosition(x, props);
-
-    if (left.wasOutsideBounds) {
-      newPosition.horizontal = TooltipHorizontalOffset.Right;
-    } else {
-      x = left.value;
-    }
-  }
-
-  if (newPosition.horizontal === TooltipHorizontalOffset.Right) {
-    const right = getRightPosition(x, props);
-    x = right.value;
-  }
-
-  if (newPosition.horizontal === TooltipHorizontalOffset.Center) {
-    const center = getCenterPosition(x, props);
-    x = center.value;
-  }
-
-  return {
-    x: clamp({
-      amount: x,
-      min: TOOLTIP_MARGIN,
-      max:
-        window.innerWidth -
-        props.tooltipDimensions.width -
-        TOOLTIP_MARGIN -
-        SCROLLBAR_WIDTH,
-    }),
-    y: clamp({
-      amount: y,
-      min: window.scrollY + TOOLTIP_MARGIN,
-      max:
-        window.scrollY +
-        window.innerHeight -
-        props.tooltipDimensions.height -
-        TOOLTIP_MARGIN,
-    }),
-  };
-}
+import {TOOLTIP_MARGIN} from './constants';
+import type {AlteredPositionProps} from './types';
 
 interface IsOutsideBoundsData {
   current: number;
