+++ packages/polaris-viz/src/components/TooltipWrapper/utilities/getLineChartTooltipPosition.ts
@@ -1,0 +1,61 @@
+import {clamp} from '@shopify/polaris-viz-core';
+import type {ScaleLinear} from 'd3-scale';
+
+import type {TooltipPositionParams} from '../types';
+import {TOOLTIP_POSITION_DEFAULT_RETURN} from '../constants';
+
+import {getXYFromEventType, eventPointNative} from './eventPoint';
+
+interface Props extends Omit<TooltipPositionParams, 'xScale'> {
+  xScale: ScaleLinear<number, number>;
+}
+
+export function getLineChartTooltipPosition({
+  chartBounds,
+  data,
+  event,
+  eventType,
+  index,
+  longestSeriesIndex,
+  xScale,
+}: Props) {
+  if (eventType === 'mouse') {
+    if (event == null) {
+      return TOOLTIP_POSITION_DEFAULT_RETURN;
+    }
+
+    const point = eventPointNative(event);
+
+    if (point == null || xScale == null || data[longestSeriesIndex] == null) {
+      return TOOLTIP_POSITION_DEFAULT_RETURN;
+    }
+
+    const {svgX} = point;
+
+    const closestIndex = Math.round(xScale.invert(svgX - chartBounds.y));
+
+    const activeIndex = clamp({
+      amount: closestIndex,
+      min: 0,
+      max: data[longestSeriesIndex].data.length - 1,
+    });
+
+    const {x, y} = getXYFromEventType(event);
+
+    return {
+      x,
+      y,
+      activeIndex,
+    };
+  } else {
+    const activeIndex = index ?? 0;
+
+    const x = xScale?.(activeIndex) ?? 0;
+
+    return {
+      x: x + chartBounds.x,
+      y: chartBounds.y,
+      activeIndex,
+    };
+  }
+}
