+++ packages/polaris-viz/src/components/VerticalBarChart/index.ts
@@ -2,3 +2,4 @@ export {VerticalBarChart} from './VerticalBarChart';
 export type {VerticalBarChartProps} from './VerticalBarChart';
 export {useVerticalBarChart} from './hooks/useVerticalBarChart';
 export {VerticalBarGroup} from './components/VerticalBarGroup';
+export {sortBarChartData} from './utilities';
