+++ packages/polaris-viz/src/components/VerticalBarChart/utilities/index.ts
@@ -1,3 +1,4 @@
 export {formatAriaLabel} from './formatAriaLabel';
 export {getBorderRadiusForStackedValues} from './getBorderRadiusForStackedValues';
 export {getYPosition} from './getYPosition';
+export {sortBarChartData} from './sortBarChartData';
