+++ packages/polaris-viz/src/components/index.ts
@@ -24,11 +24,8 @@ export {SkipLink} from './SkipLink';
 export {VisuallyHiddenRows} from './VisuallyHiddenRows';
 export {LinePreview} from './LinePreview';
 export type {LinePreviewProps} from './LinePreview';
-export {
-  TooltipWrapper,
-  TOOLTIP_POSITION_DEFAULT_RETURN,
-} from './TooltipWrapper';
-export type {TooltipPosition, TooltipPositionParams} from './TooltipWrapper';
+export {TooltipWrapper} from './TooltipWrapper';
+export type {TooltipPosition} from './TooltipWrapper';
 export {SimpleBarChart} from './SimpleBarChart';
 export type {SimpleBarChartProps} from './SimpleBarChart';
 export {Legend} from './Legend';
