+++ packages/polaris-viz/src/styles/shared/_chart-container.scss
@@ -6,6 +6,7 @@
   font-family: $font-stack-base;
   font-size: 12px;
 
+  /* stylelint-disable */
   -webkit-touch-callout: none;
   -webkit-user-select: none;
   -khtml-user-select: none;
@@ -13,4 +14,5 @@
   -ms-user-select: none;
   user-select: none;
   -webkit-tap-highlight-color: rgba(0, 0, 0, 0);
+  /* stylelint-enable */
 }
