--- packages/polaris-viz/src/utilities/isMouseEvent.ts
@@ -1,7 +1,0 @@
-import type React from 'react';
-
-export function isMouseEvent(
-  event: React.SyntheticEvent,
-): event is React.MouseEvent {
-  return event.nativeEvent instanceof MouseEvent;
-}
