--- packages/polaris-viz/src/utilities/isTouchEvent.ts
@@ -1,7 +1,0 @@
-import type React from 'react';
-
-export function isTouchEvent(
-  event: React.SyntheticEvent,
-): event is React.TouchEvent {
-  return event.nativeEvent instanceof TouchEvent;
-}
